Add solar covers MV-UV (#111)

+++ src/main/java/gregtech/common/covers/CoverSolarPanel.java
@@ -23,9 +23,9 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class CoverSolarPanel extends CoverBehavior implements ITickable {
 
-    private final int EUt;
+    private final long EUt;
 
-    public CoverSolarPanel(ICoverable coverHolder, EnumFacing attachedSide, int EUt) {
+    public CoverSolarPanel(ICoverable coverHolder, EnumFacing attachedSide, long EUt) {
         super(coverHolder, attachedSide);
         this.EUt = EUt;
     }
